refactor: use kebab-case in gatsby folder
