+++ gatsby/create-redirects.js
